refactor: next-app
