+++ packages/react-instantsearch/examples/next-app/components/instantsearch.js
@@ -1,34 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import {
-  RefinementList,
-  SearchBox,
-  Hits,
-  Configure,
-} from 'react-instantsearch/dom';
-import { InstantSearch } from './instantsearchserver';
-export default class extends React.Component {
-  static propTypes = {
-    searchState: PropTypes.object,
-    resultsState: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
-    onSearchStateChange: PropTypes.func,
-  };
+import { createInstantSearch } from 'react-instantsearch/server';
 
-  render() {
-    return (
-      <InstantSearch
-        appId="latency"
-        apiKey="6be0576ff61c053d5f9a3225e2a90f76"
-        indexName="ikea"
-        resultsState={this.props.resultsState}
-        onSearchStateChange={this.props.onSearchStateChange}
-        searchState={this.props.searchState}
-      >
-        <Configure hitsPerPage={3} />
-        <SearchBox />
-        <Hits />
-        <RefinementList attributeName="category" />
-      </InstantSearch>
-    );
-  }
-}
+const { InstantSearch, findResultsState } = createInstantSearch();
+
+export { InstantSearch, findResultsState };
