--- packages/react-instantsearch/examples/next-app/components/instantsearchserver.js
@@ -1,5 +1,0 @@
-import { createInstantSearch } from 'react-instantsearch/server';
-
-const { InstantSearch, findResultsState } = createInstantSearch();
-
-export { InstantSearch, findResultsState };
